Fix: statistics 덮어쓰기 오류 수정
