+++ backend/src/loaders/routerLoader.ts
@@ -4,6 +4,6 @@ import { useExpressServer } from 'routing-controllers';
 export default function routerLoader(app: express.Application): void {
   useExpressServer(app, {
     routePrefix: '/v1',
-    controllers: ['src/routers/v1/*.ts'],
+    controllers: ['src/routers/v1/**/*.ts'],
   });
 }
