+++ backend/src/routers/v1/users/index.ts
@@ -20,9 +20,6 @@ import {
   TistoryService,
   NotifyService,
   FollowService,
-  DashboardRepoService,
-  DashboardHistoryService,
-  ProblemService,
 } from 'src/services';
 import { Authorization, ERROR, RESPONSECODE, Cursor } from 'src/utils';
 
@@ -86,20 +83,6 @@ export default class UsersRouter {
     };
     delete result.githubUsername;
     return response.json({ code: RESPONSECODE.SUCCESS, data: result });
-  }
-
-  @Get('/dashboard')
-  async getDashboard(@Req() request: Request, @Res() response: Response) {
-    const { username } = request.query;
-    if (!username) {
-      throw new Error(ERROR.WRONG_QUERY_TYPE);
-    }
-    const dashboard = await UserService.findOneUserDashboard({ username });
-    const dashboardHistories = await DashboardHistoryService.findDashboardHistory(dashboard._id!);
-    return response.json({
-      code: RESPONSECODE.SUCCESS,
-      data: { _id: dashboard._id, username, ...dashboard.dashboard, dashboardHistories },
-    });
   }
 
   @Get('/:userID/posts')
@@ -222,27 +205,6 @@ export default class UsersRouter {
     });
   }
 
-  @Get('/:userID/dashboard/repositories')
-  async getDashboardRepoList(@Req() request: Request, @Res() response: Response) {
-    const { userID } = request.params;
-    const result = await DashboardRepoService.readDashboardRepos(userID);
-    return response.json({ code: RESPONSECODE.SUCCESS, data: result });
-  }
-
-  @Get('/:userID/dashboard/repositories/languages')
-  async getDashboardReposLanguage(@Req() request: Request, @Res() response: Response) {
-    const { userID } = request.params;
-    const result = await DashboardRepoService.readDashboardReposLanguage(userID);
-    return response.json({ code: RESPONSECODE.SUCCESS, data: result });
-  }
-
-  @Get('/:userID/dashboard/problems/statistics')
-  async getDashboardProblemsStatistics(@Req() request: Request, @Res() response: Response) {
-    const { userID } = request.params;
-    const result = await ProblemService.findOneDashboardStatistics(userID);
-    return response.json({ code: RESPONSECODE.SUCCESS, data: result });
-  }
-
   @Post('/:userID/follows')
   @UseBefore(passport.authenticate('jwt-registered', { session: false }))
   async putUserFollows(@Req() request: Request, @Res() response: Response) {
@@ -258,38 +220,6 @@ export default class UsersRouter {
     return response.json({ code: RESPONSECODE.SUCCESS });
   }
 
-  @Post('/:userID/dashboard/repositories/:repoName')
-  @UseBefore(passport.authenticate('jwt-registered', { session: false }))
-  async postDashboardRepo(@Req() request: Request, @Res() response: Response) {
-    const { userID, repoName } = request.params;
-    const { userID: bodyUserID } = request.body;
-    if (bodyUserID !== request.user?.userID) {
-      throw new Error(ERROR.PERMISSION_DENIED);
-    }
-    if (userID !== request.user!.userID) {
-      throw new Error(ERROR.WRONG_PARAMS_TYPE);
-    }
-    const githubUsername = await UserService.findGithubUsernameForUserID(userID);
-    const repoData = await GitService.findRepo(githubUsername as string, repoName);
-    const result = await DashboardRepoService.createDashboardRepo(userID, repoData);
-    return response.json({ code: RESPONSECODE.SUCCESS, data: result });
-  }
-
-  @Post('/:userID/dashboard/histories')
-  @UseBefore(passport.authenticate('jwt-registered', { session: false }))
-  async postDashboardHistory(@Req() request: Request, @Res() response: Response) {
-    const { userID } = request.params;
-    const { content, date } = request.body;
-    if (userID !== request.user!.userID) {
-      throw new Error(ERROR.WRONG_PARAMS_TYPE);
-    }
-    if (!content || !date) {
-      throw new Error(ERROR.WRONG_BODY_TYPE);
-    }
-    const history = await DashboardHistoryService.createDashboardHistory(userID, content, date);
-    return response.json({ code: RESPONSECODE.SUCCESS, data: history });
-  }
-
   @Put('/:userID/settings')
   @UseBefore(passport.authenticate('jwt', { session: false }))
   async putUser(@Req() request: Request, @Res() response: Response) {
@@ -299,47 +229,6 @@ export default class UsersRouter {
     }
     await UserService.updateOneUserConfig(request.user!.userID, request.body);
     return response.json({ code: RESPONSECODE.SUCCESS });
-  }
-
-  @Put('/:userID/dashboard')
-  @UseBefore(passport.authenticate('jwt', { session: false }))
-  async putDashboard(@Req() request: Request, @Res() response: Response) {
-    const { userID } = request.params;
-    const data = request.body;
-    if (userID !== request.user!.userID) {
-      throw new Error(ERROR.PERMISSION_DENIED);
-    }
-
-    await UserService.updateOneUserDashboard(userID, data);
-
-    const result = await UserService.findOneUserDashboard({ _id: userID });
-    return response.json({
-      code: RESPONSECODE.SUCCESS,
-      data: { _id: result._id, ...result.dashboard },
-    });
-  }
-
-  @Put('/:userID/dashboard/problems/:username/statistics')
-  @UseBefore(passport.authenticate('jwt', { session: false }))
-  async putStatistics(@Req() request: Request, @Res() response: Response) {
-    const { username, userID } = request.params;
-    if (userID !== request.user!.userID) {
-      throw new Error(ERROR.PERMISSION_DENIED);
-    }
-    const statistics = await ProblemService.findSolvedAcStatistics(username);
-    UserService.updateOneProblemStatistics(userID, statistics);
-    return response.json({ code: RESPONSECODE.SUCCESS, data: statistics });
-  }
-
-  @Put('/:userID/dashboard/repositories/languages')
-  @UseBefore(passport.authenticate('jwt-registered', { session: false }))
-  async putDashboardReposLanguage(@Req() request: Request, @Res() response: Response) {
-    const { userID } = request.params;
-    if (userID !== request.user!.userID) {
-      throw new Error(ERROR.PERMISSION_DENIED);
-    }
-    const result = await DashboardRepoService.updateDashboardReposLanguage(userID);
-    return response.json({ code: RESPONSECODE.SUCCESS, data: result });
   }
 
   @Delete('/:userID/follows')
@@ -354,32 +243,6 @@ export default class UsersRouter {
       throw new Error(ERROR.WRONG_PARAMS_TYPE);
     }
     await FollowService.removeFollow(request.user!.userID, userID);
-    return response.json({ code: RESPONSECODE.SUCCESS });
-  }
-
-  @Delete('/:userID/dashboard/histories')
-  @UseBefore(passport.authenticate('jwt-registered', { session: false }))
-  async deleteUserDashboardHistory(@Req() request: Request, @Res() response: Response) {
-    const { userID } = request.params;
-    const { historyID } = request.body;
-    if (userID !== request.user?.userID) {
-      throw new Error(ERROR.PERMISSION_DENIED);
-    }
-    if (!historyID) {
-      throw new Error(ERROR.WRONG_BODY_TYPE);
-    }
-    await DashboardHistoryService.deleteDashboardHistory(userID, historyID);
-    return response.json({ code: RESPONSECODE.SUCCESS });
-  }
-
-  @Delete('/:userID/dashboard/repositories/:repoName')
-  @UseBefore(passport.authenticate('jwt-registered', { session: false }))
-  async deleteDashboardRepo(@Req() request: Request, @Res() response: Response) {
-    const { userID, repoName } = request.params;
-    if (userID !== request.user?.userID) {
-      throw new Error(ERROR.PERMISSION_DENIED);
-    }
-    await DashboardRepoService.deleteDashboardRepo(userID, repoName);
     return response.json({ code: RESPONSECODE.SUCCESS });
   }
 }
