+++ backend/src/types/index.ts
@@ -12,4 +12,8 @@ export type ObjectType<T> = {
   [K in keyof T]: T[K];
 };
 
+export type StringKeyObject = {
+  [K: string]: object | string;
+};
+
 export * from './modelType';
